Added a service for saving any image.
