+++ Pagyeonja.API/Controllers/RiderRegistrationController.cs
@@ -7,197 +7,179 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Pagyeonja.Entities.Entities;
+using Pagyeonja.Repositories.Repositories;
 using Pagyeonja.Services.Services;
 
 namespace pagyeonjaAPI.Controllers
 {
-    [Route("api/[controller]")]
-    [ApiController]
-    public class RiderRegistrationController : ControllerBase
-    {
-        private readonly HitchContext _context;
-        private readonly RiderService _riderService;
+	[Route("api/[controller]")]
+	[ApiController]
+	public class RiderRegistrationController : ControllerBase
+	{
+		private readonly HitchContext _context;
+		private readonly RiderService _riderService;
 
-        public RiderRegistrationController(HitchContext context, RiderService riderService)
-        {
-            _context = context;
-            _riderService = riderService;
-        }
+		public RiderRegistrationController(HitchContext context, RiderService riderService)
+		{
+			_context = context;
+			_riderService = riderService;
+		}
 
-        // GET: api/Riders
-        [HttpGet("GetRiders")]
-        public async Task<ActionResult<IEnumerable<Rider>>> GetRiders()
-        {
-            try
-            {
-                return await _context.Riders.OrderByDescending(a => a.DateApplied).ToListAsync();
-            }
-            catch (Exception ex)
-            {
-                return NotFound(ex);
-            }
-        }
+		// GET: api/Riders
+		[HttpGet("GetRiders")]
+		public async Task<ActionResult<IEnumerable<Rider>>> GetRiders()
+		{
+			try
+			{
+				return await _context.Riders.OrderByDescending(a => a.DateApplied).ToListAsync();
+			}
+			catch (Exception ex)
+			{
+				return NotFound(ex);
+			}
+		}
 
-        [HttpGet("GetRidersApproved")]
-        public async Task<ActionResult<IEnumerable<Rider>>> GetRidersApproved()
-        {
-            if (_context.Riders == null)
-            {
-                return NotFound();
-            }
+		[HttpGet("GetRidersApproved")]
+		public async Task<ActionResult<IEnumerable<Rider>>> GetRidersApproved()
+		{
+			if (_context.Riders == null)
+			{
+				return NotFound();
+			}
 
-            return await _context.Riders.Where(a => a.ApprovalStatus == true).OrderByDescending(a => a.DateApplied).ToListAsync();
-        }
+			return await _context.Riders.Where(a => a.ApprovalStatus == true).OrderByDescending(a => a.DateApplied).ToListAsync();
+		}
 
 
-        // GET: api/Rider/5
-        [HttpGet("GetRider")]
-        public async Task<ActionResult<Rider>> GetRider(int id)
-        {
-            if (_context.Riders == null)
-            {
-                return NotFound();
-            }
-            var Rider = await _context.Riders.FindAsync(id);
+		// GET: api/Rider/5
+		[HttpGet("GetRider")]
+		public async Task<ActionResult<Rider>> GetRider(int id)
+		{
+			if (_context.Riders == null)
+			{
+				return NotFound();
+			}
+			var Rider = await _context.Riders.FindAsync(id);
 
-            if (Rider == null)
-            {
-                return NotFound();
-            }
+			if (Rider == null)
+			{
+				return NotFound();
+			}
 
-            return Rider;
-        }
+			return Rider;
+		}
 
-        // PUT: api/Rider/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("UpdateRider")]
-        public async Task<IActionResult> PutRider(Guid id, Rider Rider)
-        {
-            if (id != Rider.RiderId)
-            {
-                return BadRequest();
-            }
+		// PUT: api/Rider/5
+		// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+		[HttpPut("UpdateRider")]
+		public async Task<IActionResult> PutRider(Guid id, Rider Rider)
+		{
+			if (id != Rider.RiderId)
+			{
+				return BadRequest();
+			}
 
-            _context.Entry(Rider).State = EntityState.Modified;
+			_context.Entry(Rider).State = EntityState.Modified;
 
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!RiderExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+			try
+			{
+				await _context.SaveChangesAsync();
+			}
+			catch (DbUpdateConcurrencyException)
+			{
+				if (!RiderExists(id))
+				{
+					return NotFound();
+				}
+				else
+				{
+					throw;
+				}
+			}
 
-            return NoContent();
-        }
+			return NoContent();
+		}
 
-        // POST: api/Rider
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost("RegisterRider")]
-        public async Task<ActionResult<Rider>> PostRider([FromForm] Rider rider)
-        {
-            try
-            {
-                // generate riderid for rider
+		// POST: api/Rider
+		// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+		[HttpPost("RegisterRider")]
+		public async Task<ActionResult<Rider>> PostRider([FromForm] Rider rider)
+		{
+			try
+			{
+				await ((IRiderService)_riderService).AddRider(rider);
+				return CreatedAtAction("PostRider", new { id = rider.RiderId }, rider);
+			}
+			catch (Exception)
+			{
+				// Log the exception here
+				return StatusCode(500, "An error occurred while processing your request.");
+			}
+		}
 
-                await _riderService.AddRider(rider);
-                try
-                {
-                    await _context.SaveChangesAsync();
-                }
-                catch (Exception)
-                {
-                    // Log the exception here
-                    return StatusCode(500, "An error occurred while saving to the database.");
-                }
-
-                return CreatedAtAction("PostRider", new { id = rider.RiderId }, rider);
-            }
-            catch (Exception)
-            {
-                // Log the exception here
-                return StatusCode(500, "An error occurred while processing your request.");
-            }
-        }
-
-        private static async Task<List<string>> SaveImages(List<IFormFile> images)
-        {
-            var filePaths = new List<string>();
-            foreach (var image in images)
-            {
-                // Generate a unique filename
-                var extension = Path.GetExtension(image.FileName);
-                var uniqueFileName = $"{Guid.NewGuid()}{extension}";
-
-                // Save the image to the Images folder
-                var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\img", "rider_profile", uniqueFileName);
-                using var stream = new FileStream(path, FileMode.Create);
-                await image.CopyToAsync(stream);
-                filePaths.Add(uniqueFileName);
-            }
-            return filePaths;
-        }
+		[HttpPost]
+		public async Task SaveImages(Guid id, List<IFormFile> images, string doctype, string usertype)
+		{
+			try
+			{
+				await _riderService.SaveImage(id, images, doctype, usertype);
+			}
+			catch (Exception ex)
+			{
+				NotFound(ex);
+			}
+		}
 
 
 
 
-        // DELETE: api/Rider/5
-        [HttpDelete("DeleteRider")]
-        public async Task<IActionResult> DeleteRider(int id)
-        {
-            if (_context.Riders == null)
-            {
-                return NotFound();
-            }
-            var Rider = await _context.Riders.FindAsync(id);
-            if (Rider == null)
-            {
-                return NotFound();
-            }
+		// DELETE: api/Rider/5
+		[HttpDelete("DeleteRider")]
+		public async Task<IActionResult> DeleteRider(int id)
+		{
+			if (_context.Riders == null)
+			{
+				return NotFound();
+			}
+			var Rider = await _context.Riders.FindAsync(id);
+			if (Rider == null)
+			{
+				return NotFound();
+			}
 
-            _context.Riders.Remove(Rider);
-            await _context.SaveChangesAsync();
+			_context.Riders.Remove(Rider);
+			await _context.SaveChangesAsync();
 
-            return NoContent();
-        }
+			return NoContent();
+		}
 
-        // Business logic controllers
-        [HttpPut("RiderApprovalResponse")]
-        public async Task<IActionResult> RiderApprovalResponse(Guid riderId, bool response)
-        {
-            try
-            {
-                var rider = await _context.Riders.FirstOrDefaultAsync(r => r.RiderId == riderId);
-                var approval = await _context.Approvals.FirstOrDefaultAsync(a => a.UserId == riderId);
-                if (approval != null && rider != null)
-                {
-                    rider.ApprovalStatus = response;
-                    approval.ApprovalStatus = response;
-                    approval.ApprovalDate = new DateTime();
-                    await _context.SaveChangesAsync();
-                    return new JsonResult("Rider approved");
-                }
-                return new BadRequestObjectResult("Rider not found!");
-            }
-            catch (Exception ex)
-            {
-                return NotFound(ex);
-            }
-        }
+		// Business logic controllers
+		[HttpPut("RiderApprovalResponse")]
+		public async Task<IActionResult> RiderApprovalResponse(Guid riderId, bool response)
+		{
+			try
+			{
+				var rider = await _context.Riders.FirstOrDefaultAsync(r => r.RiderId == riderId);
+				var approval = await _context.Approvals.FirstOrDefaultAsync(a => a.UserId == riderId);
+				if (approval != null && rider != null)
+				{
+					rider.ApprovalStatus = response;
+					approval.ApprovalStatus = response;
+					approval.ApprovalDate = new DateTime();
+					await _context.SaveChangesAsync();
+					return new JsonResult("Rider approved");
+				}
+				return new BadRequestObjectResult("Rider not found!");
+			}
+			catch (Exception ex)
+			{
+				return NotFound(ex);
+			}
+		}
 
-        private bool RiderExists(Guid id)
-        {
-            return (_context.Riders?.Any(e => e.RiderId == id)).GetValueOrDefault();
-        }
-
-
-    }
+		private bool RiderExists(Guid id)
+		{
+			return (_context.Riders?.Any(e => e.RiderId == id)).GetValueOrDefault();
+		}
+	}
 }
