+++ Pagyeonja.Services/Pagyeonja.Services.csproj
@@ -5,6 +5,10 @@
     <ProjectReference Include="..\Pagyeonja.Entities\Pagyeonja.Entities.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.0" />
+  </ItemGroup>
+
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
