+++ Pagyeonja.Services/Services/IRiderService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Pagyeonja.Entities.Entities;
+using Microsoft.AspNetCore.Http;
 
 namespace Pagyeonja.Services.Services
 {
@@ -15,5 +16,6 @@ namespace Pagyeonja.Services.Services
         Task<Rider> AddRider(Rider rider);
         Task<bool> DeleteRider(Guid id);
         Task<bool> RiderExists(Guid id);
+        Task SaveImage(Guid id, List<IFormFile> images, string doctype, string usertype);
     }
 }
